move value_to_decimal to helper module
