+++ lib/money.rb
@@ -1,7 +1,6 @@
-require_relative 'bigdecimal'
-require_relative 'bigdecimal/util'
 require_relative 'money/money_parser'
 require_relative 'money/money'
+require_relative 'money/helpers'
 require_relative 'money/currency'
 require_relative "money/deprecations"
 require_relative 'money/accounting_money_parser'
